# Refactor ClusterMap