--- a/proto/quilkin/config/v1alpha1/config.proto
+++ b/proto/quilkin/config/v1alpha1/config.proto
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+syntax = "proto3";
+
+import "google/protobuf/struct.proto";
+
+package quilkin.config.v1alpha1;
+
+message ClusterMap {
+    repeated Cluster clusters = 1;
+}
+
+message Cluster {
+    Locality locality = 1;
+    repeated Endpoint endpoints = 2;
+}
+
+message Locality {
+    string region = 1;
+    string zone = 2;
+    string sub_zone = 3;
+}
+
+message Endpoint {
+    string host = 1;
+    uint32 port = 2;
+    google.protobuf.Struct metadata = 3;
+}