This PR refactors our internal storage of transmission of clusters over xDS to be more focused on endpoints rather than mimicking Envoys layout. We weren't using basically any of it, and having all of those extra fields was creating a lot of extra bytes and made the encoding logic more complicated.

This PR moves to using a simpler map of locality -> endpoints, as that was the only part we were using, and uses our own proto types for xDS discovery. We still use the core discovery request and response wrapper, it's just the types inside are now Quilkin types rather than envoy types.

We don't have a xDS benchmark, but I wrote a small micro benchmark of just the time and size of discovery responses, and there's a significant improvement to using this implementation, being roughly 30%-250% faster depending on the endpoint amounts (being faster for smaller amounts of endpoints where the message size overhead is more noticeable), and being 12% smaller.

### Old
endpoints | time (seconds) | size (bytes)
-|-|-
0 | 0.001 | 53
100 | 0.005 | 5537
1000 | 0.028 | 55013
10000 | 0.117 | 550013
65535 | 0.721 | 3653593

### New
endpoints | time (seconds) | size (bytes)
-|-|-
0 | 0.001 | 53
100 | 0.002 | 4910
1000 | 0.011 | 48984
10000 | 0.085 | 489984
65535 | 0.539 | 3260352



### Diff
endpoints | time (faster) | size (smaller)
-|-|-
100 | 2.5x | 12%
1000 | 2.5x | 12%
10000 | 37% | 12%
65535 | 33% | 12%

```rust
#[test]
fn protobuf_test() {
    let measure = |count: u16| {
        let config = Config::default();
        if count != 0 {
            config.clusters.write().insert_default(crate::endpoint::LocalityEndpoints::from((0..count).map(|i| Endpoint::from((std::net::Ipv4Addr::UNSPECIFIED, i))).collect::<Vec<_>>()));
        }
        let response_creation = std::time::Instant::now();
        let response = config.discovery_request("", ResourceType::Cluster, &[]).unwrap();
        let time = response_creation.elapsed().as_secs_f64();
        let size = crate::prost::encode(&response).unwrap().len();
        println!("count: {count}, time: {time}, size: {size}");
    };

    (measure)(0);
    (measure)(100);
    (measure)(1000);
    (measure)(10000);
    (measure)(u16::MAX);
    panic!();
}
```