@markmandel 2023-09-12
Lemme know if you want me to take a look, but also reminding you that we have:

https://github.com/googleforgames/quilkin/blob/d8e93b60ab0be42d8ae7b020d21302345d6dfe7b/agones/src/lib.rs#L418-L425

In case you want to try debugging in here.