@markmandel 2023-09-19
Ah, `parse_ipv6_endpoint` has `localities` in the config still

https://github.com/googleforgames/quilkin/blob/de4988956bdf27e6277b449324599b112fb86cb7/src/config.rs#L412-L424